+++ src/python/dxpy/utils/describe.py
@@ -776,7 +776,7 @@ def printable_ssh_host_key(ssh_host_key):
         return stdout.replace(" no comment", "").strip()
 
 
-def print_execution_desc(desc):
+def print_execution_desc(desc, verbose=False):
     recognized_fields = ['id', 'class', 'project', 'workspace', 'region',
                          'app', 'applet', 'executable', 'workflow',
                          'state',
@@ -789,10 +789,16 @@ def print_execution_desc(desc):
                          'delayWorkspaceDestruction', 'stages', 'totalPrice', 'isFree', 'invoiceMetadata',
                          'priority', 'sshHostKey', 'internetUsageIPs', 'spotWaitTime', 'maxTreeSpotWaitTime',
                          'maxJobSpotWaitTime', 'spotCostSavings', 'preserveJobOutputs', 'treeTurnaroundTime',
-                         'selectedTreeTurnaroundTimeThreshold', 'selectedTreeTurnaroundTimeThresholdFrom']
+                         'selectedTreeTurnaroundTimeThreshold', 'selectedTreeTurnaroundTimeThresholdFrom',
+                         'runSystemRequirements', 'runSystemRequirementsByExecutable', 'mergedSystemRequirementsByExecutable', 'runStageSystemRequirements']
 
     print_field("ID", desc["id"])
     print_field("Class", desc["class"])
+
+    if desc['class'] == 'analysis' and verbose:
+        default_analysis_desc = dxpy.DXAnalysis(desc['id']).describe()
+        desc.update(default_analysis_desc)
+
     if "name" in desc and desc['name'] is not None:
         print_field("Job name", desc['name'])
     if "executableName" in desc and desc['executableName'] is not None:
@@ -888,8 +894,6 @@ def print_execution_desc(desc):
         print_field("Failure message", desc["failureMessage"])
     if "failureFrom" in desc and desc['failureFrom'] is not None and desc['failureFrom']['id'] != desc['id']:
         print_field("Failure is from", desc['failureFrom']['id'])
-    if 'systemRequirements' in desc:
-        print_json_field("Sys Requirements", desc['systemRequirements'])
     if "tags" in desc:
         print_list_field("Tags", desc["tags"])
     if "properties" in desc:
@@ -947,6 +951,16 @@ def print_execution_desc(desc):
     if 'selectedTreeTurnaroundTimeThresholdFrom' in desc:
         print_field("Selected tree TAT from", desc['selectedTreeTurnaroundTimeThresholdFrom'])
 
+    if 'systemRequirements' in desc:
+        print_json_field("Sys Requirements", desc['systemRequirements'])
+    if 'runSystemRequirements'  in desc:
+        print_json_field("Run Sys Reqs", desc['runSystemRequirements'])
+    if 'runSystemRequirementsByExecutable' in desc:
+        print_json_field("Run Sys Reqs by Exec", desc['runSystemRequirementsByExecutable'])
+    if 'mergedSystemRequirementsByExecutable' in desc:
+        print_json_field("Merged Sys Reqs By Exec", desc['mergedSystemRequirementsByExecutable'])
+    if 'runStageSystemRequirements' in desc:
+        print_json_field("Run Stage Sys Reqs", desc['runStageSystemRequirements'])
 
     for field in desc:
         if field not in recognized_fields:
@@ -1036,6 +1050,11 @@ def print_desc(desc, verbose=False):
     Depending on the class of the entity, this method will print a
     formatted and human-readable string containing the data in *desc*.
     '''
+    if isinstance(desc, dict) and not desc.get('class'):
+        from ..utils.resolver import is_hashid
+        if is_hashid(desc.get('id')):
+            desc['class'] = desc['id'].split("-")[0]
+
     if desc['class'] in ['project', 'workspace', 'container']:
         print_project_desc(desc, verbose=verbose)
     elif desc['class'] == 'app':
@@ -1043,7 +1062,7 @@ def print_desc(desc, verbose=False):
     elif desc['class'] == 'globalworkflow':
         print_globalworkflow_desc(desc, verbose=verbose)
     elif desc['class'] in ['job', 'analysis']:
-        print_execution_desc(desc)
+        print_execution_desc(desc, verbose=verbose)
     elif desc['class'] == 'user':
         print_user_desc(desc)
     elif desc['class'] in ['org', 'team']: